+++ insights/dashboards/viewsets.py
@@ -57,6 +57,9 @@ class DashboardViewSet(
             "monitoring_average_time_metrics",
             "monitoring_peaks_in_human_service",
             "monitoring_csat_score_by_agents",
+            "finished",
+            "analysis_finished_rooms_status",
+            "analysis_peaks_in_human_service",
         ]:
             return [
                 IsAuthenticated(),
@@ -258,7 +261,8 @@ class DashboardViewSet(
     def monitoring_list_status(self, request, pk=None):
         dashboard = self.get_object()
         service = HumanSupportDashboardService(project=dashboard.project)
-        data = service.get_attendance_status(filters=request.query_params)
+        filters = {key: value for key, value in request.query_params.items()}
+        data = service.get_attendance_status(filters=filters)
         return Response(data, status=status.HTTP_200_OK)
 
     @action(
@@ -269,7 +273,8 @@ class DashboardViewSet(
     def monitoring_average_time_metrics(self, request, pk=None):
         dashboard = self.get_object()
         service = HumanSupportDashboardService(project=dashboard.project)
-        data = service.get_time_metrics(filters=request.query_params)
+        filters = {key: value for key, value in request.query_params.items()}
+        data = service.get_time_metrics(filters=filters)
         return Response(data, status=status.HTTP_200_OK)
 
     @action(
@@ -395,5 +400,42 @@ class DashboardViewSet(
     def monitoring_peaks_in_human_service(self, request, pk=None):
         dashboard = self.get_object()
         service = HumanSupportDashboardService(project=dashboard.project)
-        results = service.get_peaks_in_human_service(filters=request.query_params)
+        filters = {key: value for key, value in request.query_params.items()}
+        results = service.get_peaks_in_human_service(filters=filters)
+        return Response({"results": results}, status=status.HTTP_200_OK)
+
+    @action(
+        detail=True,
+        methods=["get"],
+        url_path="finished",
+    )
+    def finished(self, request, pk=None):
+        dashboard = self.get_object()
+        service = HumanSupportDashboardService(project=dashboard.project)
+        filters = {key: value for key, value in request.query_params.items()}
+        data = service.get_finished_rooms(filters=filters)
+        return Response(data, status=status.HTTP_200_OK)
+
+    @action(
+        detail=True,
+        methods=["get"],
+        url_path="analysis/finished_rooms_status",
+    )
+    def analysis_finished_rooms_status(self, request, pk=None):
+        dashboard = self.get_object()
+        service = HumanSupportDashboardService(project=dashboard.project)
+        filters = {key: value for key, value in request.query_params.items()}
+        data = service.get_analysis_status(filters=filters)
+        return Response(data, status=status.HTTP_200_OK)
+
+    @action(
+        detail=True,
+        methods=["get"],
+        url_path="analysis/peaks_in_human_service",
+    )
+    def analysis_peaks_in_human_service(self, request, pk=None):
+        dashboard = self.get_object()
+        service = HumanSupportDashboardService(project=dashboard.project)
+        filters = {key: value for key, value in request.query_params.items()}
+        results = service.get_analysis_peaks_in_human_service(filters=filters)
         return Response({"results": results}, status=status.HTTP_200_OK)